+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/types/EmbeddedRealmObject.kt
@@ -29,7 +29,7 @@ import io.realm.kotlin.internal.RealmObjectInternal
  * - They cannot have fields annotated with `@PrimaryKey`.
  * - When a parent object is deleted, all embedded objects are also deleted.
  */
-public interface EmbeddedRealmObject : BaseRealmObject
+public interface EmbeddedRealmObject : TypedRealmObject
 
 /**
  * TODO Put this in BaseRealmObject and just throw for DynamicRealmObject? Right now we duplicate this between RealmObject and EmbeddedRealmObject
@@ -40,3 +40,4 @@ public inline fun <reified T : EmbeddedRealmObject> T.copyFromRealm(depth: Int =
     val realm = obj.io_realm_kotlin_objectReference!!.owner.owner as TypedRealm
     return realm.copyFromRealm(obj, depth, closeAfterCopy) as T
 }
+