+++ CHANGELOG.md
@@ -6,10 +6,14 @@
 ### Enhancements
 * Add support for `Realm.copyFromRealm()`. All RealmObjects, RealmResults, RealmList and RealmSets now also have a `copyFromRealm()` extension method.
 * Add support for in-memory Realms.
+* Added support for reverse relationships through the `linkingObjects` delegate. See the function documentation for more details. (Issue [#1021](https://github.com/realm/realm-kotlin/pull/1021))
+* Added support for `BsonObjectId` and its typealias `org.mongodb.kbson.ObjectId` as a replacement for `ObjectId`. `io.realm.kotlin.types.ObjectId` is still functional but has been marked as deprecated.
 * [Sync] Added support for API key authentication. (Issue [#432](https://github.com/realm/realm-kotlin/issues/432))
+* [Sync] Added support for `BsonObjectId` as partition value.
 
 ### Fixed
-* `RealmUUID` did not calculate the correct `hashCode`, so putting it in a `HashSet` resulted in missed values.
+* `RealmUUID` did not calculate the correct `hashCode`, so putting it in a `HashSet` resulted in duplicates.
+* Close underlying realm if it is no longer referenced by any Kotlin object. (Issue [#671](https://github.com/realm/realm-kotlin/issues/671))
 
 ### Compatibility
 * This release is compatible with the following Kotlin releases:
@@ -23,6 +27,7 @@
 * Minimum Android SDK: 16.
 
 ### Internal
+* Added dependency Kbson 0.1.0.
 * Updated to use hierarchical multi platform project structure.
 * Updated to Realm Core 12.10.0, commit 8ce82fe3a8d5a2fbc89d719de8559f5a792c2dc9.